refactor(strategy): 패키지 변경
